fix: guard unix ioctl behind build constraint
golang.org/x/sys/unix does not compile on Windows, causing
build failures on non-Unix platforms. Split fromTTY() into
a build-constrained file (termwidth_unix.go, !windows) and
provide a no-op stub (termwidth_windows.go) so the package
compiles everywhere; Detect() falls back to $COLUMNS / 120
on platforms without /dev/tty.
