+++ cmd/swm/internal/termwidth/termwidth.go
@@ -4,14 +4,12 @@ package termwidth
 import (
 	"os"
 	"strconv"
-
-	"golang.org/x/sys/unix"
 )
 
 const defaultWidth = 120
 
 // Detect returns the terminal width using the following fallback chain:
-//  1. term.GetSize on /dev/tty (if the file can be opened and returns width > 0)
+//  1. /dev/tty ioctl (Unix only, via fromTTY)
 //  2. $COLUMNS environment variable parsed as a positive integer
 //  3. defaultWidth (120)
 func Detect() int {
@@ -30,22 +28,6 @@ func DetectFromEnv() int {
 	}
 
 	return defaultWidth
-}
-
-func fromTTY() int {
-	f, err := os.Open("/dev/tty")
-	if err != nil {
-		return 0
-	}
-
-	defer f.Close() //nolint:errcheck // closing a read-only device file cannot fail meaningfully
-
-	ws, err := unix.IoctlGetWinsize(int(f.Fd()), unix.TIOCGWINSZ)
-	if err != nil {
-		return 0
-	}
-
-	return int(ws.Col)
 }
 
 func fromColumns() int {
