+++ cmd/swm/internal/termwidth/termwidth_windows.go
@@ -1,0 +1,7 @@
+//go:build windows
+
+package termwidth
+
+func fromTTY() int {
+	return 0
+}
